+++ chapter 2/3-final_app/pages/unit_converter.py
@@ -1,5 +1,7 @@
 import taipy.gui.builder as tgb
-from taipy.gui import notify
+from food_converter_functions.change_volume_callback import change_volume
+from food_converter_functions.change_weight_callback import change_weight
+from food_converter_functions.reset_callback import reset
 
 ###########################
 ## Initial values        ##
@@ -8,48 +10,9 @@ grams = cups = ounces = tablespoons = teaspoons = milliliters = 0
 
 
 ###########################
-## Callbacks             ##
+## Page                  ##
 ###########################
-def change_weight(state, var_name, value):
-    if var_name == "grams":  # Grams to Ounces
-        state.ounces = round(value / 28.35, 2)
-    else:  # Ounces to Grams
-        state.grams = round(value * 28.35, 2)
 
-
-def change_volume(state, var_name, value):
-    if var_name == "cups":
-        # state.tablespoons = round(value * 16, 2)
-        state.tablespoons = value * 16
-        state.milliliters = round(value * 236.6, 2)
-        state.teaspoons = round(value * 48, 2)
-    elif var_name == "tablespoons":
-        state.cups = round(value / 16, 2)
-        state.milliliters = round(value * 14.787, 2)
-        state.teaspoons = round(value * 3, 2)
-    elif var_name == "milliliters":
-        state.tablespoons = round(value / 14.787, 2)
-        state.cups = round(value / 236.6, 2)
-        state.teaspoons = round(value / 4.929, 2)
-    elif var_name == "teaspoons":
-        state.milliliters = round(value * 4.929, 2)
-        state.cups = round(value / 48, 2)
-        state.tablespoons = round(value / 3, 2)
-
-
-def reset(state, id, payload):
-    if id != "reset_volume":
-        state.grams = 0
-        state.ounces = 0
-    if id != "reset_weight":
-        state.cups = 0
-        state.tablespoons = 0
-        state.teaspoons = 0
-        state.milliliters = 0
-    notify(state, notification_type="warning", message="Values set to 0")
-
-
-lov_test = [True, False]
 with tgb.Page() as converter_page:  # food_fact_page for food_facts.py
     tgb.text("# Unit converter 🧮", mode="md")
